requirements.txt
Flask
pymongo
dnspython
python-dotenv
PyJWT